Added username field to create_user layout Finsihed API for create_user Re-Organized code base
