+++ app/src/main/java/API/CreateUserAPI.java
@@ -1,17 +1,22 @@
-package softeng.eventplanning;
+package API;
 
 import android.os.AsyncTask;
 import android.util.Log;
 
 import org.json.JSONObject;
 
+import java.io.BufferedInputStream;
+import java.io.BufferedReader;
 import java.io.DataOutputStream;
+import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.io.OutputStreamWriter;
+import java.io.StringWriter;
 import java.net.HttpURLConnection;
 import java.net.URL;
 import java.net.URLEncoder;
 
-public class CreateAPI extends AsyncTask<String,String,String> {
+public class CreateUserAPI extends AsyncTask<String,String,String> {
     private  String[] marray;
     @Override
     protected void onPreExecute() {
@@ -24,7 +29,7 @@ public class CreateAPI extends AsyncTask<String,String,String> {
 
     @Override
     protected String doInBackground(String ... params) {
-        String urlstring = new String("http://10.0.2.2:5000/");
+        String urlstring = new String("http://10.0.2.2:5000/create-user");
         DataOutputStream printout;
         JSONObject jsonobj = new JSONObject();
 
@@ -33,15 +38,41 @@ public class CreateAPI extends AsyncTask<String,String,String> {
 
             jsonobj.put("email",marray[0]);
             jsonobj.put("username",marray[1]);
-            jsonobj.put("email",marray[2]);
+            jsonobj.put("password",marray[2]);
+
+            String urlparam = new String();
+            urlparam = jsonobj.toString();
+
+            System.out.println(urlparam);
+
             URL url = new URL(urlstring);
             HttpURLConnection connection = (HttpURLConnection) url.openConnection();
             connection.setRequestMethod("POST");
-            printout = new DataOutputStream(connection.getOutputStream ());
+            connection.setRequestProperty("User-Agent","Mozilla/5.0");
+            connection.setRequestProperty("Accept-Language","en-US,en;q=0.5");
+            connection.setDoInput(true);
 
-            printout.writeBytes(URLEncoder.encode(jsonobj.toString(),"UTF-8"));
+            printout = new DataOutputStream(connection.getOutputStream());
+
+            printout.writeBytes(urlparam);
             printout.flush ();
             printout.close ();
+
+            InputStream in = new BufferedInputStream(connection.getInputStream());
+            BufferedReader reader = new BufferedReader(new InputStreamReader(in));
+
+            StringBuilder out = new StringBuilder();
+            String line = new String();
+
+            while((line = reader.readLine()) != null){
+                out.append(line);
+            }
+
+            System.out.println(out.toString());
+            reader.close();
+
+
+
             String printy;
             Log.d("completed", printy= "POSted");
