+++ app/src/main/res/layout/create_user.xml
@@ -1,7 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:orientation="vertical" android:layout_width="match_parent"
-    android:layout_height="match_parent">
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
 
     <RelativeLayout
         android:layout_width="match_parent"
@@ -9,79 +10,91 @@
         android:layout_gravity="center_horizontal">
 
         <TextView
+            android:id="@+id/textView11"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:textAppearance="?android:attr/textAppearanceLarge"
-            android:text="Create New User"
-            android:id="@+id/textView11"
             android:layout_alignParentTop="true"
             android:layout_centerHorizontal="true"
-            android:layout_marginTop="45dp" />
+            android:layout_marginTop="45dp"
+            android:text="Create New User"
+            android:textAppearance="?android:attr/textAppearanceLarge" />
 
         <EditText
+            android:id="@+id/email"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:inputType="textEmailAddress"
-            android:ems="10"
-            android:id="@+id/editText"
+            android:layout_alignEnd="@+id/password"
+            android:layout_alignLeft="@+id/password"
+            android:layout_alignRight="@+id/password"
+            android:layout_alignStart="@+id/password"
             android:layout_below="@+id/textView11"
             android:layout_marginTop="57dp"
-            android:text="email address"
-            android:layout_alignRight="@+id/editText2"
-            android:layout_alignEnd="@+id/editText2"
-            android:layout_alignLeft="@+id/editText2"
-            android:layout_alignStart="@+id/editText2" />
+            android:ems="10"
+            android:inputType="textEmailAddress"
+            android:text="email address" />
 
         <TextView
+            android:id="@+id/text12"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:textAppearance="?android:attr/textAppearanceMedium"
             android:text="password"
-            android:id="@+id/textView12"
-            android:layout_below="@+id/editText"
-            android:layout_alignLeft="@+id/editText"
-            android:layout_alignStart="@+id/editText"
-            android:layout_marginTop="43dp" />
+            android:textAppearance="?android:attr/textAppearanceMedium"
+            android:layout_centerVertical="true"
+            android:layout_alignLeft="@+id/password"
+            android:layout_alignStart="@+id/password"
+            android:layout_below="@+id/username"
+            android:layout_marginTop="30dp"/>
 
         <EditText
+            android:id="@+id/password"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:inputType="textPassword"
             android:ems="10"
-            android:id="@+id/editText2"
-            android:layout_below="@+id/textView12"
+            android:inputType="textPassword"
+            android:layout_below="@+id/text12"
             android:layout_centerHorizontal="true" />
 
         <TextView
+            android:id="@+id/textView13"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:textAppearance="?android:attr/textAppearanceMedium"
             android:text="re-enter password"
-            android:id="@+id/textView13"
-            android:layout_below="@+id/editText2"
-            android:layout_alignLeft="@+id/editText2"
-            android:layout_alignStart="@+id/editText2"
-            android:layout_marginTop="77dp" />
+            android:textAppearance="?android:attr/textAppearanceMedium"
+            android:layout_below="@+id/password"
+            android:layout_alignLeft="@+id/password"
+            android:layout_alignStart="@+id/password"
+            android:layout_marginTop="30dp"/>
 
         <EditText
+            android:id="@+id/passwordConf"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:inputType="textPassword"
             android:ems="10"
-            android:id="@+id/editText3"
+            android:inputType="textPassword"
             android:layout_below="@+id/textView13"
             android:layout_alignLeft="@+id/textView13"
             android:layout_alignStart="@+id/textView13" />
 
         <Button
+            android:id="@+id/button"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="MAKE ME"
-            android:id="@+id/button"
-            android:layout_marginTop="64dp"
-            android:layout_below="@+id/editText3"
+            android:layout_marginTop="30dp"
+            android:layout_below="@+id/passwordConf"
             android:layout_centerHorizontal="true"
-            android:clickable="true" />
+            android:clickable="true"
+            android:text="Create User" />
+
+        <EditText
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:id="@+id/username"
+            android:text="Username"
+            android:ems="10"
+            android:layout_alignStart="@+id/email"
+            android:layout_alignLeft="@+id/email"
+            android:layout_below="@+id/email"
+            android:layout_marginTop="30dp"/>
 
     </RelativeLayout>
 </LinearLayout>
